Adding terms and conditions text
